% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illa.R
\name{check_decreasing}
\alias{check_decreasing}
\title{Determine if biomarker decreases with age and perform monotonicity checks}
\usage{
check_decreasing(df)
}
\arguments{
\item{df}{A \code{tibble} with the following columns: \code{subid}, \code{age}, \code{val}
\itemize{
\item \code{subid} is a numeric subject identifier
\item \code{age} is age in years at each observation (cannot be repeated within
subject)
\item \code{value} is the biomarker value to be modeled
}}
}
\value{
TRUE if biomarker decreases with age overall
}
\description{
Determine if biomarker decreases with age and perform monotonicity checks
}
