% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illa.R
\name{illa}
\alias{illa}
\title{Estimate value vs. time curve from longitudinal data}
\usage{
illa(df, val0)
}
\arguments{
\item{df}{A \code{tibble} with the following columns: \code{subid}, \code{age}, \code{val}
\itemize{
\item \code{subid} is a numeric subject identifier
\item \code{age} is age in years at each observation (cannot be repeated within
subject)
\item \code{value} is the biomarker value to be modeled
}}

\item{val0}{Anchor biomarker value that corresponds to \code{t = 0}}
}
\value{
\code{illa} returns an object of class "sila" containg:
\itemize{
\item \code{val0} Anchor biomarker value that corresponds to \code{t = 0}
\item \code{adtime_to_val} function representing g : val = g(adtime)
\item \code{val_to_rate} function representing f : rate = f(val)
\item \code{tdrs} a \code{tibble} with sampled rates at query points
\item \code{is_decreasing} TRUE if biomarker value globally decreases with age
}
}
\description{
Estimate value vs. time curve from longitudinal data
}
\examples{
set.seed(42)
df <- tibble::tibble(
  subid = c(
    1, 1, 1,
    2, 2, 2, 2, 2, 2
  ),
  age = c(
    seq(from = 50, to = 70, length.out = 3),
    seq(from = 50, to = 70, length.out = 6)
  ),
  val = c(
    2, 4, 6,
    2, 4, 6, 8, 10, 12
  ) + stats::rnorm(3 + 6, mean = 0, sd = .1)
)
illa(df, val0 = 2)
}
