% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sila_estimate.R
\name{sila_estimate}
\alias{sila_estimate}
\title{Generate subject-level estimates of time to threshold based on the modeled
value vs. time function}
\usage{
sila_estimate(
  sila_res,
  df = NULL,
  align_event = "all",
  adtime_limits = c(-50, 50)
)
}
\arguments{
\item{sila_res}{output of illa, an object of class "sila"}

\item{df}{A \code{tibble} with the following columns: \code{subid}, \code{age}, \code{val}
\itemize{
\item \code{subid} is a numeric subject identifier
\item \code{age} is age in years at each observation (cannot be repeated within
subject)
\item \code{value} is the biomarker value to be modeled
}}

\item{align_event}{One of: "all", "first", "last", "mean". "all" uses all
visits per subject to determine their time shifts. "first" uses the first
visit per subject only. "last" uses the last visit per subject only. "mean"
uses the mean value and mean age within each subject to estimate the shift.}

\item{adtime_limits}{a vector with two elements specifying the min and max
allowed values for adjusted time}
}
\value{
A tibble containing the following columns:
\itemize{
\item \code{subid}: subject identifier
\item \code{age}: age in years at observation
\item \code{val}: observed biomarker value
\item \code{shift}: estimated time shift
\item \code{adtime}: estimated adjusted time, given by age + shift
}
}
\description{
Generate subject-level estimates of time to threshold based on the modeled
value vs. time function
}
\examples{
set.seed(42)
df <- tibble::tibble(
  subid = c(
    1, 1, 1,
    2, 2, 2, 2, 2, 2
  ),
  age = c(
    seq(from = 50, to = 70, length.out = 3),
    seq(from = 50, to = 70, length.out = 6)
  ),
  val = c(
    2, 4, 6,
    2, 4, 6, 8, 10, 12
  ) + stats::rnorm(3 + 6, mean = 0, sd = .1)
)
res <- illa(df, val0 = 2)
sila_estimate(res)
}
