% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulated_longitudinal_data}
\alias{simulated_longitudinal_data}
\title{Simulated longitudinal data.}
\format{
\subsection{\code{simulated_longitudinal_data}}{

A data frame with 600 rows and 3 columns:
\describe{
\item{subid}{Subject identifier}
\item{age}{Age at visit}
\item{val}{Biomarker value at visit}
}
}
}
\usage{
simulated_longitudinal_data
}
\description{
Simulated longitudinal data.
}
\keyword{datasets}
