% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illa.R
\name{illa}
\alias{illa}
\title{Estimate value vs. time curve from longitudinal data.}
\usage{
illa(df, dt, val0, maxi, skern)
}
\arguments{
\item{df}{A tibble with the following columns: subid, age, val}

\item{dt}{Time difference between queries}

\item{val0}{Anchor value}

\item{maxi}{Maximum number of iterations}

\item{skern}{Span of the smoothing kernel}
}
\value{
A list containing two tibbles
}
\description{
Estimate value vs. time curve from longitudinal data.
}
