% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illa.R
\name{illa}
\alias{illa}
\title{Estimate value vs. time curve from longitudinal data}
\usage{
illa(df, dt, val0, maxi, skern)
}
\arguments{
\item{df}{A tibble with the following columns: subid, age, val}

\item{dt}{Time difference between queries}

\item{val0}{Anchor value}

\item{maxi}{Maximum number of iterations}

\item{skern}{Span of the smoothing kernel}
}
\value{
A list containing two tibbles
}
\description{
Estimate value vs. time curve from longitudinal data
}
\examples{
set.seed(42)
df <- tibble::tibble(
  subid = c(
    1, 1, 1,
    2, 2, 2, 2, 2, 2
  ),
  age = c(
    seq(from = 50, to = 70, length.out = 3),
    seq(from = 50, to = 70, length.out = 6)
  ),
  val = c(
    2, 4, 6,
    2, 4, 6, 8, 10, 12
  ) + stats::rnorm(3 + 6, mean = 0, sd = .1)
)
illa(df, dt = 2, val0 = 2, maxi = 100, skern = 0)
illa(df, dt = 2, val0 = 2, maxi = 100, skern = 0.2)
}
