% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illa.R
\name{illa}
\alias{illa}
\title{Estimate value vs. time curve from longitudinal data}
\usage{
illa(df, dt, val0, maxi, skern)
}
\arguments{
\item{df}{A \code{tibble} with the following columns: \code{subid}, \code{age}, \code{val}.
\itemize{
\item \code{subid} is a numeric subject identifier
\item \code{age} is age in years at each observation (cannot be repeated within
subject)
\item \code{value} is the biomarker value to be modeled
}}

\item{dt}{Time difference between queries (numerical integration step size)}

\item{val0}{Anchor value that corresponds to \code{t = 0}}

\item{maxi}{Maximum number of iterations}

\item{skern}{Span of the smoothing kernel}
}
\value{
A list containing two \code{tibble}s
}
\description{
Estimate value vs. time curve from longitudinal data
}
\examples{
set.seed(42)
df <- tibble::tibble(
  subid = c(
    1, 1, 1,
    2, 2, 2, 2, 2, 2
  ),
  age = c(
    seq(from = 50, to = 70, length.out = 3),
    seq(from = 50, to = 70, length.out = 6)
  ),
  val = c(
    2, 4, 6,
    2, 4, 6, 8, 10, 12
  ) + stats::rnorm(3 + 6, mean = 0, sd = .1)
)
illa(df, dt = 2, val0 = 2, maxi = 100, skern = 0)
illa(df, dt = 2, val0 = 2, maxi = 100, skern = 0.2)
}
