% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sila.R
\name{sila}
\alias{sila}
\title{Sampled Iterative Local Approximation with smoothing}
\usage{
sila(df, dt, val0, maxi, sk = NULL)
}
\arguments{
\item{df}{A tibble with the following columns: subid, age, val.
subid is a numeric subject identifier, age is age in years at each
observation, and value is the biomarker value to be modeled.
Age cannot be repeated within subject.}

\item{dt}{Numerical integration step size}

\item{val0}{Anchor value that corresponds to t = 0}

\item{maxi}{Maximum number of iterations}

\item{sk}{List of smoothing kernels}
}
\value{
A list containing two tibbles:
- tsila specifies the value vs. time curve: val is the SILA-modeled
value, time is the SILA-modeled time resulting from numeric
integration (not anchored), adtime is the anchored time such that
t = 0 corresponds to val0, mrate is the mean sampled rate through the
value, sdrate is the mean sampled rate through the value, nsubs is
the number of subjects with observations that intersect the modeled
value, sdval is an approximation of the standard deviation of the
value (calculated by propagating the rate error), ci95 is an
approximation of the 95\% confidence interval of the value (likely
underestimated since it does not account for temporal covariance
across observations)
- tdrs
}
\description{
Sampled Iterative Local Approximation with smoothing
}
\examples{
set.seed(42)
df <- tibble::tibble(
  subid = c(
    1, 1, 1,
    2, 2, 2, 2, 2, 2
  ),
  age = c(
    seq(from = 50, to = 70, length.out = 3),
    seq(from = 50, to = 70, length.out = 6)
  ),
  val = c(
    2, 4, 6,
    2, 4, 6, 8, 10, 12
  ) + stats::rnorm(3 + 6, mean = 0, sd = .1)
)
sila_res <- sila(df, dt = 2, val0 = 2, maxi = 100)
}
