% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sila_estimate.R
\name{sila_estimate}
\alias{sila_estimate}
\title{Generate subject-level estimates of time to threshold based on the modeled
value vs. time function}
\usage{
sila_estimate(
  tsila,
  df,
  align_event = "last",
  extrap_years = 3,
  truncate_aget0 = TRUE
)
}
\arguments{
\item{tsila}{Tibble output of SILA containing the modeled value vs. time
function}

\item{df}{A tibble with the following columns: subid, age, val}

\item{align_event}{Which observation(s) to use within a subject to estimate
subject-level age of onset and duration of positivity.
Can be "first", "last" (default), or "all".}

\item{extrap_years}{Data within extrap_years is used to extrapolate SILA
estimates when observations fall beyond the modeled
range. Linear regression is applied to the modeled
relationship between value and time. Default is 3.}

\item{truncate_aget0}{Boolean (default is TRUE) specifying if subject-level
estimates should be truncated such that the lowest
duration estimate is set to the earliest modeled time
point on the SILA curve.}
}
\value{
A tibble containing the following columns:
- subid: subject identifier
- age: age in years at observation
- val: observed biomarker value
- minage: minimum observed age of subject
- maxage: maximum observed age of subject
- valt0: time point defining zero time
- ageref: age at the reference observation
- dtageref: time from observation to reference
- estval: SILA-estimated value at observation
- estage0: SILA-estimated age the subject will cross the threshold
- estdtt0: SILA-estimated time from threshold
- estresid: Estimated biomarker residual
- estpos: Boolean indicating if SILA estimate is above threshold
- aevent: Same as align_event
- extrapyrs: Same as extrap_years
- truncated: Boolean indicating if estaget0 and estdtt0 are truncated
}
\description{
Generate subject-level estimates of time to threshold based on the modeled
value vs. time function
}
\examples{
set.seed(42)
df <- tibble::tibble(
  subid = c(
    1, 1, 1,
    2, 2, 2, 2, 2, 2
  ),
  age = c(
    seq(from = 50, to = 70, length.out = 3),
    seq(from = 50, to = 70, length.out = 6)
  ),
  val = c(
    2, 4, 6,
    2, 4, 6, 8, 10, 12
  ) + stats::rnorm(3 + 6, mean = 0, sd = .1)
)
illa_res <- illa(df, dt = 2, val0 = 2, maxi = 100, skern = 0)
sila_estimate(illa_res$tout, df)
sila_estimate(illa_res$tout, df, align_event = "first")
sila_estimate(illa_res$tout, df, align_event = "all")
}
